Moved gpu wrapping stuff to own file
